+++ CHANGELOG.md
@@ -3,7 +3,17 @@ This files contains all the documented changes in this project. The project adhe
 
 
 
+
+[1.0.0] - 2016-07-14
+#added
+Added write.createImg()
+Added helped function to read files, parse bitmap headers, and transform bitmap pallets
+
+
+
+
 [0.1.0] - 2016-07-14
 #added
 Added project skeleton
+